requirements.txt
absl-py==2.2.0
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
altair==5.5.0
annotated-types==0.7.0
anyascii==0.3.2
anyio==4.8.0
asgiref==3.8.1
asttokens==3.0.0
attrs==25.1.0
audioread==3.0.1
babel==2.17.0
backoff==2.2.1
bangla==0.0.2
bcrypt==4.3.0
beautifulsoup4==4.13.3
blinker==1.9.0
blis==1.2.0
bnnumerizer==0.0.2
bnunicodenormalizer==0.1.7
build==1.2.2.post1
cachetools==5.5.1
catalogue==2.0.10
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.1
chroma-hnswlib==0.7.6
chromadb==0.6.3
click==8.1.8
cloudpathlib==0.21.0
colorama==0.4.6
coloredlogs==15.0.1
confection==0.1.5
contourpy==1.3.1
coqpit==0.0.17
cycler==0.12.1
cymem==2.0.11
Cython==3.0.12
dataclasses-json==0.6.7
dateparser==1.1.8
decorator==5.2.1
Deprecated==1.2.18
distro==1.9.0
docopt==0.6.2
durationpy==0.9
einops==0.8.1
encodec==0.1.1
executing==2.2.0
faiss-cpu==1.10.0
fastapi==0.115.11
filelock==3.17.0
Flask==3.1.0
Flask-Cors==5.0.0
flatbuffers==25.2.10
fonttools==4.56.0
frozenlist==1.5.0
fsspec==2025.3.0
g2pkk==0.1.2
gitdb==4.0.12
GitPython==3.1.44
google-auth==2.38.0
googleapis-common-protos==1.69.2
greenlet==3.1.1
groq==0.15.0
grpcio==1.71.0
gruut==2.2.3
gruut-ipa==0.13.0
gruut_lang_de==2.0.1
gruut_lang_en==2.0.1
gruut_lang_es==2.0.1
gruut_lang_fr==2.0.2
h11==0.14.0
hangul-romanize==0.1.0
httpcore==1.0.7
httptools==0.6.4
httpx==0.28.1
httpx-sse==0.4.0
huggingface-hub==0.29.3
humanfriendly==10.0
icecream==2.1.4
idna==3.10
importlib_metadata==8.6.1
importlib_resources==6.5.2
inflect==7.5.0
itsdangerous==2.2.0
jamo==0.4.1
jieba==0.42.1
Jinja2==3.1.5
jiter==0.8.2
joblib==1.4.2
jsonlines==1.2.0
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
kiwisolver==1.4.8
kubernetes==32.0.1
langchain==0.3.16
langchain-chroma==0.2.2
langchain-community==0.3.16
langchain-core==0.3.47
langchain-groq==0.2.3
langchain-ollama==0.3.0
langchain-openai==0.3.2
langchain-text-splitters==0.3.5
langcodes==3.5.0
langsmith==0.3.2
language_data==1.3.0
lazy_loader==0.4
librosa==0.11.0
llvmlite==0.44.0
lxml==5.3.1
marisa-trie==1.2.1
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.26.0
matplotlib==3.10.1
mdurl==0.1.2
mmh3==5.1.0
monotonic==1.6
more-itertools==10.6.0
mpmath==1.3.0
msgpack==1.1.0
multidict==6.1.0
murmurhash==1.0.12
mypy-extensions==1.0.0
narwhals==1.24.0
networkx==2.8.8
nltk==3.9.1
num2words==0.5.14
numba==0.61.0
numpy==1.26.4
oauthlib==3.2.2
ollama==0.4.7
onnxruntime==1.21.0
openai==1.60.2
opentelemetry-api==1.31.1
opentelemetry-exporter-otlp-proto-common==1.31.1
opentelemetry-exporter-otlp-proto-grpc==1.31.1
opentelemetry-instrumentation==0.52b1
opentelemetry-instrumentation-asgi==0.52b1
opentelemetry-instrumentation-fastapi==0.52b1
opentelemetry-proto==1.31.1
opentelemetry-sdk==1.31.1
opentelemetry-semantic-conventions==0.52b1
opentelemetry-util-http==0.52b1
orjson==3.10.15
overrides==7.7.0
packaging==24.2
pandas==1.5.3
pillow==11.1.0
platformdirs==4.3.7
pooch==1.8.2
posthog==3.21.0
preshed==3.0.9
propcache==0.2.1
protobuf==5.29.3
psutil==7.0.0
pyarrow==19.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pydantic==2.10.6
pydantic-settings==2.7.1
pydantic_core==2.27.2
pydeck==0.9.1
Pygments==2.19.1
pynndescent==0.5.13
pyparsing==3.2.1
PyPika==0.48.9
pypinyin==0.53.0
pyproject_hooks==1.2.0
pyreadline3==3.5.4
pysbd==0.3.4
python-crfsuite==0.9.11
python-dateutil==2.9.0.post0
python-docx==1.1.2
python-dotenv==1.0.1
pytz==2024.2
PyYAML==6.0.2
RapidFuzz==3.12.2
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rich==13.9.4
rpds-py==0.22.3
rsa==4.9
safetensors==0.5.3
scikit-learn==1.6.1
scipy==1.15.1
sentence-transformers==3.4.1
sentencepiece==0.2.0
shellingham==1.5.4
six==1.17.0
smart-open==7.1.0
smmap==5.0.2
sniffio==1.3.1
soundfile==0.13.1
soupsieve==2.6
soxr==0.5.0.post1
spacy==3.8.4
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SQLAlchemy==2.0.37
srsly==2.5.1
starlette==0.46.1
streamlit==1.41.1
SudachiDict-core==20250129
SudachiPy==0.6.10
sympy==1.13.1
tenacity==9.0.0
tensorboard==2.19.0
tensorboard-data-server==0.7.2
thinc==8.3.4
threadpoolctl==3.5.0
tiktoken==0.8.0
tokenizers==0.21.0
toml==0.10.2
torch==2.6.0
torchaudio==2.6.0
tornado==6.4.2
tqdm==4.67.1
trainer==0.0.36
transformers==4.49.0
typeguard==4.4.2
typer==0.15.2
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2025.1
tzlocal==5.3.1
umap-learn==0.5.7
Unidecode==1.3.8
urllib3==2.3.0
uvicorn==0.34.0
wasabi==1.1.3
watchdog==6.0.0
watchfiles==1.0.4
weasel==0.4.1
websocket-client==1.8.0
websockets==15.0.1
Werkzeug==3.1.3
wrapt==1.17.2
yarl==1.18.3
zipp==3.21.0
zstandard==0.23.0
